# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с фильмами",
        "title": "MovieHub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/movies/list": {
            "get": {
                "description": "Возвращает все фильмы из БД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Получить список фильмов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie_short"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movies/search": {
            "get": {
                "description": "Возвращает список фильмов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Поиск фильма по названию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie title",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie_short"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Возвращает полную информацию о фильме",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Поиск фильма по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Movie_ex"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Cast": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "enName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                }
            }
        },
        "model.ExternalId": {
            "type": "object",
            "properties": {
                "tmdb": {
                    "type": "integer"
                }
            }
        },
        "model.Movie_ex": {
            "type": "object",
            "properties": {
                "awards": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "budget": {
                    "type": "integer"
                },
                "cast": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Cast"
                    }
                },
                "country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "enName": {
                    "type": "string"
                },
                "externalId": {
                    "$ref": "#/definitions/model.ExternalId"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "movieType": {
                    "type": "string"
                },
                "movieYear": {
                    "type": "integer"
                },
                "poster": {
                    "type": "string"
                },
                "ratings": {
                    "$ref": "#/definitions/model.Ratings"
                },
                "revenue": {
                    "type": "integer"
                },
                "runame": {
                    "type": "string"
                },
                "seasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Seasons"
                    }
                },
                "source": {
                    "$ref": "#/definitions/model.Source"
                },
                "top250": {
                    "type": "integer"
                }
            }
        },
        "model.Movie_short": {
            "type": "object",
            "properties": {
                "enName": {
                    "type": "string"
                },
                "externalId": {
                    "$ref": "#/definitions/model.ExternalId"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "movieType": {
                    "type": "string"
                },
                "movieYear": {
                    "type": "integer"
                },
                "ratings": {
                    "$ref": "#/definitions/model.Ratings"
                },
                "runame": {
                    "type": "string"
                }
            }
        },
        "model.Ratings": {
            "type": "object",
            "properties": {
                "filmCritic": {
                    "type": "number",
                    "format": "float64"
                },
                "kp": {
                    "type": "number",
                    "format": "float64"
                },
                "russianFilmCritics": {
                    "type": "number",
                    "format": "float64"
                },
                "tmdb": {
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "model.Seasons": {
            "type": "object",
            "properties": {
                "episodesCount": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
        "model.Source": {
            "type": "object",
            "properties": {
                "kphd": {
                    "type": "boolean"
                },
                "tmdb": {
                    "type": "boolean"
                }
            }
        }
    }
}